+++ CHANGELOG.md
@@ -13,6 +13,7 @@ being manually edited by hand.
     - Side Bar background: #002138 -> #001E33
     - Status Bar background: #001E33 -> #002540
     - Current line: #003154 -> #002C4C
+- Changed Waves pink: #FF80DD -> #F77FBE
 - Booleans and numbers are now *italic*.
 - Properties now use the tag color. (Red in variants.)
 - Escape characters now use the function color. (Blue in variants.)